+++ src/create-block.js
@@ -3,7 +3,7 @@ const d3 = require('d3');
 class CreateBlock {
 
 	constructor() {
-		// this.checkanswer = this.checkanswer.bind(this)
+
 	}
 
 	start() {
@@ -20,8 +20,7 @@ class CreateBlock {
 
 		var alphabet = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z"];
 
-		var everything = d3.select("#vis").append("div").attr('class', 'everything')
-		var block_container = everything.append("div").attr('class', 'blockContainer')
+		var block_container = d3.select("#vis").append("div").attr('class', 'blockContainer')
 
 		var grid = block_container.append('div').attr('class', 'grid');
 
@@ -38,28 +37,34 @@ class CreateBlock {
 
 		var encryption = text_container.append('div').attr('class', 'encryption');
 		encryption.append('div').attr('class', 'block_message').text("");
-		encryption.append('div').attr('class', 'cipher_message').text("");		
+		encryption.append('div').attr('class', 'cipher_message').text("");
+		
 
-		var interactive_container = everything.append("div").attr('class', 'interactiveContainer')
+		// this.message_transition();
 
-		// var interactive_container = d3.select('.interactiveContainer')
-		// interactive_container.append('h2').text('Test your block cipher skills!')
-		// interactive_container.append('h3').text('Choose a word from the dropdown below, and try to cipher it. When you are done click on the check button to see if you got it right :)')
-		// var textboxes = interactive_container.append('div').attr('class', 'textboxes')
+	}
 
-		// var words = ['CHOOSE A WORD', 'BASE', 'LOVE', 'GENDER', 'FINGER', 'ACADEMIC', 'ACTIVIST']
-		// var ciphered = ['CHOOSE A WORD', 'CBUC', 'MPZA', 'KBOCBU', 'GKMHBU', 'BDBEBPHD', 'BDYOYFTU']
-		// var dropdown = textboxes.append('select').attr('id', 'dropdown')
+	message_transition() {
+		var initialMessage = "SECRET";
+		var cipherMessage = "UCBSDU"
+		var delay = 1000
+		var duration = 1000
+		var alphabet = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z"];
 
-		// for (var i = 0; i < words.length; i++) {
-		// 	dropdown.append('option').text(words[i])
-		// }
+		for(var i=0; i < initialMessage.length; i++){
+			d3.select('.block_message').append("div").attr('id', "encr_block_"+i).attr('class', 'block_msg_letter')
+				.transition()
+				.duration(duration)
+				.text(initialMessage[(i == 0) ? 0 : i])
+				.delay(delay);
 
-		// textboxes.append('input').style('margin-left', '20px').attr("id", 'userinput')
+			d3.select('.cipher_message').append("div").attr('id', "decr_block_"+i).attr('class', 'block_msg_letter update_margin')
+				.transition()
+				.duration(duration)
+				.delay(delay);
+		}
 
-		// textboxes.append('div')
-		// 	.attr('id', 'checkbutton')
-		// 	.text('check')
+		var last = 0
 
 		for(var i=0; i < initialMessage.length; i+=2){
 			d3.select("#encr_block_" + i)
@@ -95,8 +100,37 @@ class CreateBlock {
 					.style('background-color', '#e53935')
 					.delay((last * 2000)+(delay + duration));
 
-	transitions(plain_text_pairs, cipher_text_pairs, i) {
-		var rule = (i==0) ? "column" : "row"; 
+				d3.select('.block-' + alphabet[index + 5])
+					.transition()
+					.duration(duration)
+					.style('background-color', '#ffcccc')
+					.delay((last * 2000)+(delay + duration));
+				d3.select('.block-' + alphabet[index + 6])
+					.transition()
+					.duration(duration)
+					.style('background-color', '#ffcccc')
+					.delay((last * 2000)+(delay + duration));
+				d3.select('.block-' + alphabet[index + 7])
+					.transition()
+					.duration(duration)
+					.style('background-color', '#ffcccc')
+					.delay((last * 2000)+(delay + duration));
+				
+				d3.select('.block-' + alphabet[index + 10])
+					.transition()
+					.duration(duration)
+					.style('background-color', '#ffcccc')
+					.delay((last * 2000)+(delay + duration));
+				d3.select('.block-' + alphabet[index + 11])
+					.transition()
+					.duration(duration)
+					.style('background-color', '#ffcccc')
+					.delay((last * 2000)+(delay + duration));
+				d3.select('.block-' + alphabet[index + 12])
+					.transition()
+					.duration(duration)
+					.style('background-color', '#ffcccc')
+					.delay((last * 2000)+(delay + duration));
 
 				d3.select('.block-' + alphabet[index + 15])
 					.transition()
@@ -211,20 +245,25 @@ class CreateBlock {
 							.style('background-color', 'transparent')
 							.delay(1500);
 		
-		var duration = 500;
-		var delay = 1000;
+						d3.select('.block-' + alphabet[index + 15])
+							.transition()
+							.duration(duration)
+							.style('background-color', 'transparent')
+							.delay(1500);
+						d3.select('.block-' + alphabet[index + 16])
+							.transition()
+							.duration(duration)
+							.style('background-color', 'transparent')
+							.delay(1500);
+						d3.select('.block-' + alphabet[index + 17])
+							.transition()
+							.duration(duration)
+							.style('background-color', 'transparent')
+							.delay(1500)
+					});	
 
-		// color the plain text letters in the grid (red)
-		d3.select('.block-' + plain_text_pairs[i][0])
-			.transition()
-			.duration(duration)
-			.style('background-color', 'red')
-			.delay(0);
-		d3.select('.block-' + plain_text_pairs[i][1])
-			.transition()
-			.duration(duration)
-			.style('background-color', 'red')
-			.delay(0);
+			
+			}
 
 			if(i == 2) {
 				var actual_i = i;
@@ -322,18 +361,236 @@ class CreateBlock {
 							.delay(800);
 					});		
 
+			}
+
+
+			last += 1
+
+		}
+	}
+
+
+	box_transition(plain_text_pairs, cipher_text_pairs, i) {
+		var column_ex = d3.select(".rules").append("div").attr('class', 'exampleContainer').append("h3").text( "Box rule:")
+		var plain_text = column_ex.append('div').attr('class', 'example')
+
+		plain_text.append('div').text(plain_text_pairs[i][0]).attr('class', 'block_letter')
+
+		plain_text.append('div').text(plain_text_pairs[i][1]).attr('class', 'block_letter')
+
+		var cipher_text = column_ex.append('div').attr('class', 'example')
+		cipher_text.append('div').attr('id', 'box_cipher_1').attr('class', 'block_letter')
+		cipher_text.append('div').attr('id', 'box_cipher_2').attr('class', 'block_letter')
+
+		var alphabet = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z"];
+
+		var duration = 500;
+		var delay = 1000;
+		var forreal = this;
+		// highlight row 1 (box 0 is red, everything else is pink)
+		d3.select('.block-' + plain_text_pairs[i][0])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#e53935')
+			.delay(0);
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 1)])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(0);
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 2)])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(0);
+		
+		// highlight row 2
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 5)])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(0);
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 6)])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(0);
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 7)])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(0);
+
+		// highlight row 3 (box 2 is red, everything else is pink)
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 10)])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(0);
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 11)])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(0);
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 12)])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#e53935')
+			.delay(0);
+		
+		// make everything pink
+		d3.select('.block-' + plain_text_pairs[i][0])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(delay);
+
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 12)])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(delay);
+
+		// make box 2 of row 1 and box 0 of row 2 red
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 2)])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#e53935')
+			.delay(delay + duration);
+
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 10)])
+			.transition()
+			.duration(500)
+			.style('background-color', '#e53935')
+			.delay(delay + duration);
+
+		// add in the ciphered letters at the same time as the cipher text highlight
+		d3.select("#box_cipher_1")
+			.transition()
+			.duration(duration)
+			.text(cipher_text_pairs[i][0])
+			.delay(delay + duration);
+
+		d3.select("#box_cipher_2")
+			.transition()
+			.duration(duration)
+			.text(cipher_text_pairs[i][1])
+			.delay(delay + duration);
+
+		// decolor everything
+		d3.select('.block-' + plain_text_pairs[i][0])
+			.transition()
+			.duration(duration)
+			.style('background-color', "transparent")
+			.delay(2*(delay+duration));
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 1)])
+			.transition()
+			.duration(duration)
+			.style('background-color', 'transparent')
+			.delay(2*(delay+duration));
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 2)])
+			.transition()
+			.duration(duration)
+			.style('background-color', 'transparent')
+			.delay(2*(delay+duration));
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 5)])
+			.transition()
+			.duration(duration)
+			.style('background-color', 'transparent')
+			.delay(2*(delay+duration));
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 6)])
+			.transition()
+			.duration(duration)
+			.style('background-color', 'transparent')
+			.delay(2*(delay+duration));
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 7)])
+			.transition()
+			.duration(duration)
+			.style('background-color', 'transparent')
+			.delay(2*(delay+duration));
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 10)])
+			.transition()
+			.duration(duration)
+			.style('background-color', 'transparent')
+			.delay(2*(delay+duration));
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 11)])
+			.transition()
+			.duration(duration)
+			.style('background-color', "transparent")
+			.delay(2*(delay+duration));
+		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 12)])
+			.transition()
+			.duration(duration)
+			.style('background-color', "transparent")
+			.delay(2*(delay+duration))
+			.on('end', function () {
+				forreal.message_transition();
+			});
+	}
+
+	transitions(plain_text_pairs, cipher_text_pairs, i) {
+		var rule = (i==0) ? "column" : "row"; 
+
+		var column_ex = d3.select(".rules").append("div").attr('class', 'exampleContainer').append("h3").text( rule[0].toUpperCase() + rule.substr(1, rule.length) + " rule:")
+		var plain_text = column_ex.append('div').attr('class', 'example')
+
+		plain_text.append('div').text(plain_text_pairs[i][0]).attr('class', 'block_letter')
+
+		plain_text.append('div').text(plain_text_pairs[i][1]).attr('class', 'block_letter')
+
+		var cipher_text = column_ex.append('div').attr('class', 'example')
+		cipher_text.append('div').attr('id', rule + '_cipher_1').attr('class', 'block_letter')
+		cipher_text.append('div').attr('id', rule + '_cipher_2').attr('class', 'block_letter')
+		
+		var duration = 500;
+		var delay = 1000;
+
+		// color the plain text letters in the grid (red)
+		d3.select('.block-' + plain_text_pairs[i][0])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#e53935')
+			.delay(0);
+		d3.select('.block-' + plain_text_pairs[i][1])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#e53935')
+			.delay(0);
+
+		// color the plain text letters in the grid (light red)
+		d3.select('.block-' + plain_text_pairs[i][0])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(duration + delay);
+		d3.select('.block-' + plain_text_pairs[i][1])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#ffcccc')
+			.delay(duration + delay);
+
+		// color the cipher text letters in the grid (red)
+		d3.select('.block-' + cipher_text_pairs[i][0])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#e53935')
+			.delay((duration + delay));
+		d3.select('.block-' + cipher_text_pairs[i][1])
+			.transition()
+			.duration(duration)
+			.style('background-color', '#e53935')
+			.delay((duration + delay));
+
 		// add in the ciphered letters at the same time as the cipher text highlight
 		d3.select("#"+ rule + "_cipher_1")
 			.transition()
 			.duration(duration)
-			.style('color', 'blue')
 			.text(cipher_text_pairs[i][0])
 			.delay((duration + delay) + 500);
 
 		d3.select("#"+ rule + "_cipher_2")
 			.transition()
 			.duration(duration)
-			.style('color', 'blue')
 			.text(cipher_text_pairs[i][1])
 			.delay((duration + delay) + 500);
 
@@ -380,316 +637,6 @@ class CreateBlock {
 			});
 		}
 	}
-
-	box_transition(plain_text_pairs, cipher_text_pairs, i) {
-		var column_ex = d3.select(".rules").append("div").attr('class', 'exampleContainer').append("h3").text( "Box rule:")
-		var plain_text = column_ex.append('div').attr('class', 'example')
-
-		plain_text.append('div').text(plain_text_pairs[i][0]).attr('class', 'block_letter')
-
-		plain_text.append('div').text(plain_text_pairs[i][1]).attr('class', 'block_letter')
-
-		var cipher_text = column_ex.append('div').attr('class', 'example')
-		cipher_text.append('div').attr('id', 'box_cipher_1').attr('class', 'block_letter')
-		cipher_text.append('div').attr('id', 'box_cipher_2').attr('class', 'block_letter')
-
-		var alphabet = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z"];
-
-		var duration = 500;
-		var delay = 1000;
-		var forreal = this;
-		
-		// highlight M and Y in red 
-		d3.select('.block-' + plain_text_pairs[i][0])
-			.transition()
-			.duration(duration)
-			.style('background-color', '#e53935')
-			.delay(0);
-
-		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 12)])
-			.transition()
-			.duration(duration)
-			.style('background-color', '#e53935')
-			.delay(0);
-		
-		// highlight M and Y in pink
-		d3.select('.block-' + plain_text_pairs[i][0])
-			.transition()
-			.duration(duration)
-			.style('background-color', '#ffcccc')
-			.delay(delay);
-
-		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 12)])
-			.transition()
-			.duration(duration)
-			.style('background-color', '#ffcccc')
-			.delay(delay);
-
-		// highlight O and W in red 
-		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 2)])
-			.transition()
-			.duration(duration)
-			.style('background-color', '#e53935')
-			.delay(delay + duration);
-
-		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 10)])
-			.transition()
-			.duration(500)
-			.style('background-color', '#e53935')
-			.delay(delay + duration);
-
-		// add in the ciphered letters at the same time as the cipher text highlight
-		d3.select("#box_cipher_1")
-			.transition()
-			.duration(duration)
-			.style('color', 'blue')
-			.text(cipher_text_pairs[i][0])
-			.delay(delay + duration);
-
-		d3.select("#box_cipher_2")
-			.transition()
-			.duration(duration)
-			.style('color', 'blue')
-			.text(cipher_text_pairs[i][1])
-			.delay(delay + duration);
-
-		// decolor everything
-		d3.select('.block-' + plain_text_pairs[i][0])
-			.transition()
-			.duration(duration)
-			.style('background-color', "transparent")
-			.delay(2*(delay+duration));
-		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 2)])
-			.transition()
-			.duration(duration)
-			.style('background-color', 'transparent')
-			.delay(2*(delay+duration));
-		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 10)])
-			.transition()
-			.duration(duration)
-			.style('background-color', 'transparent')
-			.delay(2*(delay+duration));
-		d3.select('.block-' + alphabet[(alphabet.indexOf(plain_text_pairs[i][0]) + 12)])
-			.transition()
-			.duration(duration)
-			.style('background-color', "transparent")
-			.delay(2*(delay+duration))
-			.on('end', function () {
-				forreal.message_transition();
-			});
-	}
-
-	message_transition() {
-		var initialMessage = "SECRET";
-		var cipherMessage = "UCBSDU"
-		var alphabet = ["A", "B", "C", "D", "E", "F", "G", "H", "I", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z"];
-		
-		var duration = 1000
-		var buffer = 2500;
-
-		for(var i=0; i < initialMessage.length; i++){
-			var delay = 1000
-			d3.select('.block_message').append("div").attr('id', "encr_block_"+i).attr('class', 'block_msg_letter')
-				.transition()
-				.duration(duration)
-				.text(initialMessage[(i == 0) ? 0 : i])
-				.delay(delay);
-
-			d3.select('.cipher_message').append("div").attr('id', "decr_block_"+i).attr('class', 'block_msg_letter update_margin')
-				.transition()
-				.duration(duration)
-				.delay(delay);
-		}
-
-		var lastTiming = 0;
-
-		for(var i=0; i < initialMessage.length; i+=2){
-			var delay = (i==0) ? 0 : (lastTiming - ((i/2) * 3000));
-
-		// color the plain text letters in the grid (red)
-		d3.select('.block-' + plain_text_pairs[i][0])
-			.transition()
-			.duration(duration)
-			.style('background-color', '#e53935')
-			.delay(0);
-		d3.select('.block-' + plain_text_pairs[i][1])
-			.transition()
-			.duration(duration)
-			.style('background-color', '#e53935')
-			.delay(0);
-
-			d3.select("#encr_block_" + (i+1))
-				.transition()
-				.duration(duration)
-				.style('color', 'red')
-				.delay((last * buffer)+(delay + duration));	
-			
-
-		// color the cipher text letters in the grid (red)
-		d3.select('.block-' + cipher_text_pairs[i][0])
-			.transition()
-			.duration(duration)
-			.style('background-color', '#e53935')
-			.delay((duration + delay));
-		d3.select('.block-' + cipher_text_pairs[i][1])
-			.transition()
-			.duration(duration)
-			.style('background-color', '#e53935')
-			.delay((duration + delay));
-
-			// make two plain text letters red 
-			d3.select('.block-' + alphabet[e_index1])
-				.transition()
-				.duration(duration)
-				.style('background-color', 'red')
-				.delay((last * buffer)+(delay + duration));
-			d3.select('.block-' + alphabet[e_index2])
-				.transition()
-				.duration(duration)
-				.style('background-color', 'red')
-				.delay((last * buffer)+(delay + duration));
-
-			delay = (i==0) ? 2000 : (lastTiming - ((i/2) * 1000) - (1000*(i-2)))
-
-			d3.select("#encr_block_" + i)
-				.transition()
-				.duration(duration)
-				.style('color', 'black')
-				.delay((last * buffer)+(delay + duration));
-
-			d3.select("#encr_block_" + (i+1))
-				.transition()
-				.duration(duration)
-				.style('color', 'black')
-				.delay((last * buffer)+(delay + duration));	
-			
-			// make two plain text letters pink
-			d3.select('.block-' + alphabet[e_index1])
-				.transition()
-				.duration(duration)
-				.style('background-color', 'pink')
-				.delay((last * buffer)+(delay + duration));
-			d3.select('.block-' + alphabet[e_index2])
-				.transition()
-				.duration(duration)
-				.style('background-color', 'pink')
-				.delay((last * buffer)+(delay + duration));
-
-			var c_index1 = alphabet.indexOf(cipherMessage[i])
-			var c_index2 = alphabet.indexOf(cipherMessage[i+1])
-			
-			// make two cipher text letters red
-			d3.select('.block-' + alphabet[c_index1])
-				.transition()
-				.duration(duration)
-				.style('background-color', 'red')
-				.delay((last * buffer)+(delay + duration));
-			d3.select('.block-' + alphabet[c_index2])
-				.transition()
-				.duration(duration)
-				.style('background-color', 'red')
-				.delay((last * buffer)+(delay + duration));
-
-			d3.select("#decr_block_" + i)
-				.transition()
-				.duration(duration)
-				.style('color', 'blue')
-				.text(cipherMessage[i])
-				.delay((last * buffer)+(delay + duration));
-
-			d3.select("#decr_block_" + (i+1))
-				.transition()
-				.duration(duration)
-				.style('color', 'blue')
-				.text(cipherMessage[i+1])
-				.delay((last * buffer)+(delay + duration));	
-
-			delay = (i==0) ? 3000 : (lastTiming - ((i-2) * 500))
-			
-			// decolor everything
-			d3.select('.block-' + alphabet[e_index1])
-				.transition()
-				.duration(duration)
-				.style('background-color', 'transparent')
-				.delay((last * buffer)+(delay + duration));
-			d3.select('.block-' + alphabet[e_index2])
-				.transition()
-				.duration(duration)
-				.style('background-color', 'transparent')
-				.delay((last * buffer)+(delay + duration));
-			d3.select('.block-' + alphabet[c_index1])
-				.transition()
-				.duration(duration)
-				.style('background-color', 'transparent')
-				.delay((last * buffer)+(delay + duration));
-			
-			var forreal = this;
-			
-			if (i == 4) {
-				d3.select('.block-' + alphabet[c_index2])
-					.transition()
-					.duration(duration)
-					.style('background-color', 'transparent')
-					.delay((last * buffer)+(delay + duration))
-					.on('end', function() {
-						var interactive_container = d3.select('.interactiveContainer')
-						interactive_container.append('h2').text('Test your block cipher skills!')
-						interactive_container.append('h3').text('Choose a word from the dropdown below, and try to cipher it. When you are done click on the check button to see if you got it right :)')
-						var textboxes = interactive_container.append('div').attr('class', 'textboxes')
-
-						var words = ['CHOOSE A WORD', 'BASE', 'LOVE', 'GENDER', 'FINGER', 'ACADEMIC', 'ACTIVIST']
-						var dropdown = textboxes.append('select').attr('id', 'dropdown')
-
-						for (var i = 0; i < words.length; i++) {
-							dropdown.append('option').text(words[i])
-						}
-
-						textboxes.append('input').style('margin-left', '20px').attr("id", 'userinput')
-
-						textboxes.append('div')
-							.attr('id', 'checkbutton')
-							.text('check')
-
-						textboxes.append('div').attr('id', 'shareResult')
-							
-						document.getElementById("checkbutton").onclick = forreal.checkanswer;
-
-					});
-			} else {
-				d3.select('.block-' + alphabet[c_index2])
-					.transition()
-					.duration(duration)
-					.style('background-color', 'transparent')
-					.delay((last * buffer)+(delay + duration));
-			}
-			
-			lastTiming = (last * buffer)+(delay + duration) + duration
-		}
-	}
-
-	checkanswer() {
-		console.log("YO");
-		var selected_word = d3.select('#dropdown option:checked').text();
-		var words = ['CHOOSE A WORD', 'BASE', 'LOVE', 'GENDER', 'FINGER', 'ACADEMIC', 'ACTIVIST']
-		var ciphered = ['CHOOSE A WORD', 'CBUC', 'MPZA', 'KBOCBU', 'GKMHBU', 'BDBEBPHD', 'BDYOYFTU']
-
-		if (selected_word != 'CHOOSE A WORD') {
-			var userinput = document.getElementById("userinput").value
-			var index = words.indexOf(selected_word);
-
-			console.log(userinput, ciphered[index])
-			
-			
-			if (userinput.toUpperCase() === ciphered[index]) {
-				console.log("CONGRATS")
-				d3.select("#shareResult").text("congrats!")
-			}else{
-				console.log("try again")
-				d3.select("#shareResult").text("try again!")
-			}
-		}
-	}
-
 }
 
 module.exports = CreateBlock;
